requirements.txt
python-telegram-bot==20.3
feedparser
Flask
openai
apscheduler
python-dotenv
requests
pillow